requirements.txt
numpy==1.23.5
Pillow==9.4.0
scikit-learn==1.1.3
fastapi==0.95.0
python-multipart==0.0.6
requests==2.28.2
tensorflow==2.11.0